* dl-symbol-exporting:
  * Undefine real headers/symbols and define our own to avoid redeclaration errors
  * Update symbols to match latest API

+++ src/libcrypto.h
@@ -1,11 +1,16 @@
 #ifndef _LIBCRYPTO_H
 #define _LIBCRYPTO_H
 
+#include ".defs/libcrypto_pre.h"
+
 #include <openssl/crypto.h>
 #include <openssl/aes.h>
 #include <openssl/blowfish.h>
 #include <openssl/md5.h>
 #include <openssl/sha.h>
+
+#include ".defs/libcrypto_post.h"
+
 #include "src/crypto/aes_util.h"
 #include "src/crypto/bf_util.h"
 #include "src/crypto/dh_util.h"
@@ -14,7 +19,7 @@
 #include "dl.h"
 #include <bdlib/src/String.h>
 
-typedef void (*AES_cbc_encrypt_t)(const unsigned char*, unsigned char*, const unsigned long, const AES_KEY*, unsigned char*, const int);
+typedef void (*AES_cbc_encrypt_t)(const unsigned char*, unsigned char*, const size_t, const AES_KEY*, unsigned char*, const int);
 typedef void (*AES_decrypt_t)(const unsigned char*, unsigned char*, const AES_KEY*);
 typedef void (*AES_encrypt_t)(const unsigned char*, unsigned char*, const AES_KEY*);
 typedef int (*AES_set_decrypt_key_t)(const unsigned char*, const int, AES_KEY*);
